+++ packages/melt/CHANGELOG.md
@@ -1,5 +1,28 @@
 # melt
 
+## 0.4.0
+
+### Minor Changes
+
+- 7b165bc: add progress builder
+
+## 0.3.0
+
+### Minor Changes
+
+- fb5b860: Add Radio Group
+
+### Patch Changes
+
+- 8375026: Update typings for selection state and tree
+- 9e42d71: fix radio group issues
+
+## 0.2.2
+
+### Patch Changes
+
+- 3860900: fix tree nav
+
 ## 0.2.1
 
 ### Patch Changes